Final changes

+++ backend_Yannis.py
@@ -13,10 +13,44 @@ from pandas import read_csv
 from datetime import datetime
 from json import load
 from urllib.error import HTTPError
+import urllib3.exceptions
+import socket
+import requests
 
+#------------------------------ TABLE FUNCTION ----------------------------
+def checkTableExists(tablename):
+    sql = """SHOW TABLES LIKE '%s'""" % \
+	(tablename)
+    cursor.execute(sql)
+    result = cursor.fetchone()
+    if result:
+        return True
+    else:
+        return False
+
+#---------------------------------------------------------------------------
+def checkEntryExists(tablename,key_name,value):
+    try:
+        value = int(value)
+        sql = """SELECT * FROM %s WHERE %s = '%d' """ % \
+        	       (tablename,key_name,value)
+    except ValueError:
+        pass
+        sql = """SELECT * FROM %s WHERE %s = '%s' """ % \
+	       (tablename,key_name,value)
+
+    cursor.execute(sql)
+    result = cursor.fetchone()
+    if result:
+        return True
+    else:
+        return False
 
 #------------------------------ ALIAS FUNCTION ----------------------------
 def CreateTableAlias():
+    if checkTableExists('alias'):
+        return
+
     # Drop table if it already exist using execute() method.
     cursor.execute("DROP TABLE IF EXISTS alias")
 
@@ -37,30 +71,37 @@ def CreateTableAlias():
         # Commit your changes in the database
         db.commit()
     except:
-        print ("CreateTableAlias:DB connection error!\n")
+        # print ("CreateTableAlias:DB connection error!\n")
         # Rollback in case there is any error
         db.rollback()
 #--------------------------------------------------------------------------
 
 def insertValueAlias(alias,twitter_id,fb_id):
-  sql = """INSERT INTO alias (AID,UID_TWITTER,UID_FB) \
+    if checkEntryExists('alias','AID',alias):
+        return
+
+    sql = """INSERT INTO alias (AID,UID_TWITTER,UID_FB) \
 	VALUES ('%s', '%d', '%d')""" % \
 	(str(alias),int(twitter_id), int(fb_id))
 
-  try:
-    # Execute the SQL command
-    cursor.execute(sql)
-    # Commit your changes in the database
-    db.commit()
-  except:
-      print ("insertValueAlias:DB connection error!\n")
-    # Rollback in case there is any error
+    try:
+        # Execute the SQL command
+        cursor.execute(sql)
+        # Commit your changes in the database
+        db.commit()
+        print("Successfuly added alias: ",alias)
+    except:
+      print ("insertValueAlias:DB error!\n",sql,"\n")
+      # Rollback in case there is any error
       db.rollback()
 
 #------------------------------ FB FUNCTIONS ------------------------------
 
 
 def CreateTableFBUsers():
+    if checkTableExists('fb_users'):
+        return
+
     # Drop table if it already exist using execute() method.
     cursor.execute("DROP TABLE IF EXISTS fb_users")
 
@@ -76,13 +117,16 @@ def CreateTableFBUsers():
         # Commit your changes in the database
         db.commit()
     except:
-        print ("CreateTableFBUsers:DB connection error!\n")
+        # print ("CreateTableFBUsers:DB connection error!\n")
         # Rollback in case there is any error
         db.rollback()
 
 #--------------------------------------------------------------------------
 
 def CreateTableFBPosts():
+    if checkTableExists('fb_posts'):
+        return
+
     # Drop table if it already exist using execute() method.
     cursor.execute("DROP TABLE IF EXISTS fb_posts")
 
@@ -105,45 +149,75 @@ FOREIGN KEY (UID_POST)
         # Commit your changes in the database
         db.commit()
     except:
-        print ("CreateTableFBPosts:DB connection error!\n")
+        # print ("CreateTableFBPosts:DB connection error!\n")
         # Rollback in case there is any error
         db.rollback()
 
 #--------------------------------------------------------------------------
 
 def insertValueFBUsers(user_id, fb_user):
-  sql = """INSERT INTO fb_users (UID, NAME) \
-	VALUES ('%d', '%s')""" % \
-	(int(user_id), str(fb_user))
-
-  try:
-    # Execute the SQL command
-    cursor.execute(sql)
-    # Commit your changes in the database
-    db.commit()
-  except:
-      print ("insertValueFBUsers:DB connection error!\n")
-    # Rollback in case there is any error
-      db.rollback()
+    if checkEntryExists('fb_users','UID',user_id):
+        return
+    else:
+        sql = """INSERT INTO fb_users (UID, NAME) \
+            VALUES ('%d', '%s')""" % \
+            (int(user_id), str(fb_user))
+        try:
+            # Execute the SQL commandv
+            cursor.execute(sql)
+            # Commit your changes in the database
+            db.commit()
+        except:
+            print ("insertValueFBUsers:DB error!\n",sql,"\n")
+            # Rollback in case there is any error
+            db.rollback()
 
 #--------------------------------------------------------------------------
 
 def insertValueFBPosts(post_id, user_id, title, time, likes, comments, shares):
-  if((post_id == 'None') or (user_id == 'None')):
-      return
-  sql = """INSERT INTO fb_posts (PID, UID_POST, TITLE, CREATED_AT, LIKES, COMMENTS, SHARES) \
-	VALUES ('%d', '%d', '%s', '%s' ,'%d', '%d', '%d')""" % \
-	(int(post_id), int(user_id), title, time, likes, comments, shares)
-
-  try:
-    # Execute the SQL command
-    cursor.execute(sql)
-    # Commit your changes in the database
-    db.commit()
-  except:
-      print ("insertValueFBPosts:DB connection error!\n")
-    # Rollback in case there is any error
-      db.rollback()
+    if((post_id == None) or (user_id == None)):
+        return 0
+    if checkEntryExists('fb_posts','PID',post_id):
+        sql = """UPDATE fb_posts SET LIKES = '%d', COMMENTS = '%d', SHARES = '%d'
+        WHERE PID = '%d'""" % \
+      	(int(likes), int(comments), int(shares),int(post_id))
+        try:
+          # Execute the SQL command
+          cursor.execute(sql)
+          # Commit your changes in the database
+          db.commit()
+          return 1
+        except:
+          print ("updateValueFBPosts:DB error!\n",sql,"\n")
+          # Rollback in case there is any error
+          db.rollback()
+          return 0
+    else:
+        if (time != None):
+            try:
+                time = time.strftime("%Y-%m-%d %H:%M:%S")    
+            except:
+                time = ''
+        sql = """INSERT INTO fb_posts (PID, UID_POST, TITLE, CREATED_AT, LIKES, COMMENTS, SHARES) \
+	           VALUES ('%d', '%d', '%s', '%s' ,'%d', '%d', '%d')""" % \
+                   (int(post_id),
+                    int(user_id),
+                    title.replace("'", " ").replace('"', ' '),
+                    time, 
+                    int(likes),
+                    int(comments), 
+                    int(shares))
+        try:
+          # Execute the SQL command
+          cursor.execute(sql)
+          # Commit your changes in the database
+          db.commit()
+          return 1
+        except:
+          print ("insertValueFBPosts:DB error!\n",sql,"\n")
+          # Rollback in case there is any error
+          db.rollback()
+          return 0
 
 #--------------------------------------------------------------------------
 
@@ -152,22 +226,26 @@ def RecordFBValuesToDB(fb_user, post_list):
         user_id = post_list[0]['user_id']
         insertValueFBUsers(user_id, fb_user)
         # key from dictionaries
+        count = 0
         for i in range (0, len(post_list)):
             post = post_list[i]
-            insertValueFBPosts(str(post['post_id']),
+            count = count + insertValueFBPosts(post['post_id'],
             user_id,
-            str(post['text']),
-            post['time'].strftime("%Y-%m-%d %H:%M:%S"),
-            int(post['likes']),
-            int(post['comments']),
-            int(post['shares']))
+            post['text'],
+            post['time'],
+            post['likes'],
+            post['comments'],
+            post['shares'])
+
+        print ('Stored successfully in DB ',count,' posts')
 
 #--------------------------------------------------------------------------
 
 def fb_scraper(fb_user):
     try:
-        get_p = facebook_scraper.get_posts(fb_user, pages=15)
-    except HTTPError as ex:
+        get_p = facebook_scraper.get_posts(fb_user, pages=150)
+    except (HTTPError, requests.exceptions.RequestException, ValueError,
+            socket.timeout,ConnectionError,urllib3.exceptions ) as ex:
         print('Failed with exception [%s]' % ex)
         get_p = []
 
@@ -182,6 +260,9 @@ def fb_scraper(fb_user):
 #------------------------------ TWITTER FUNCTIONS ----------------------------
 
 def CreateTableTwitterUsers():
+    if checkTableExists('twitter_users'):
+        return
+
     # Drop table if it already exist using execute() method.
     cursor.execute("DROP TABLE IF EXISTS twitter_users")
 
@@ -201,13 +282,16 @@ def CreateTableTwitterUsers():
         # Commit your changes in the database
         db.commit()
     except:
-        print ("CreateTableTwitterUsers:DB connection error!\n")
+        # print ("CreateTableTwitterUsers:DB connection error!\n")
         # Rollback in case there is any error
         db.rollback()
 
 #--------------------------------------------------------------------------
 
 def CreateTableTwitterPosts():
+    if checkTableExists('twitter_tweets'):
+        return
+
     # Drop table if it already exist using execute() method.
     cursor.execute("DROP TABLE IF EXISTS twitter_tweets")
 
@@ -215,7 +299,7 @@ def CreateTableTwitterPosts():
     sql = """CREATE TABLE twitter_tweets (
     TID BIGINT NOT NULL,
     UID_TWEET BIGINT NOT NULL,
-    TITLE VARCHAR(320),
+    TITLE VARCHAR(1024),
     RETWEET_COUNT INT,
     FAVORITE_COUNT INT,
     CREATED_AT DATETIME,
@@ -232,60 +316,93 @@ def CreateTableTwitterPosts():
         # Commit your changes in the database
         db.commit()
     except:
-        print ("CreateTableTwitterPosts:DB connection error!\n")
+        # print ("CreateTableTwitterPosts:DB connection error!\n")
         # Rollback in case there is any error
         db.rollback()
 
 #--------------------------------------------------------------------------
 
 def insertValueTwitterUsers(user_id, twitter_user, favourites_count, followers_count, friends_count, statuses_count):
-  sql = """INSERT INTO twitter_users (UID, NAME, FAVORITES_COUNT, FOLLOWERS_COUNT, FRIENDS_COUNT, TWEETS_COUNT) \
-	VALUES ('%d', '%s', '%d', '%d', '%d', '%d')""" % \
-	(int(user_id),
-    str(twitter_user),
-    int(favourites_count),
-    int(followers_count),
-    int(friends_count),
-    int(statuses_count))
-
-  try:
-    # Execute the SQL command
-    cursor.execute(sql)
-    # Commit your changes in the database
-    db.commit()
-  except:
-      print ("insertValueTwitterUsers:DB connection error!\n")
-    # Rollback in case there is any error
-      db.rollback()
+    if checkEntryExists('twitter_users','UID',user_id):
+        sql = """UPDATE twitter_users SET FAVORITES_COUNT = '%d', FOLLOWERS_COUNT = '%d'
+            WHERE UID = '%d'""" % \
+            (favourites_count, followers_count,int(user_id))
+        try:
+            # Execute the SQL command
+            cursor.execute(sql)
+            # Commit your changes in the database
+            db.commit()
+            return 1
+        except:
+            print ("updateValueTwitterPosts:DB error!\n",sql,"\n")
+            # Rollback in case there is any error
+            db.rollback()
+            return 0
+    else:
+        sql = """INSERT INTO twitter_users (UID, NAME, FAVORITES_COUNT, FOLLOWERS_COUNT, FRIENDS_COUNT, TWEETS_COUNT) \
+	           VALUES ('%d', '%s', '%d', '%d', '%d', '%d')""" % \
+	              (int(user_id),
+                  str(twitter_user),
+                  int(favourites_count),
+                  int(followers_count),
+                  int(friends_count),
+                  int(statuses_count))
+        try:
+            # Execute the SQL command
+            cursor.execute(sql)
+            # Commit your changes in the database
+            db.commit()
+            return 1
+        except:
+            print ("insertValueTwitterUsers:DB error!\n",sql,"\n")
+            # Rollback in case there is any error
+            db.rollback()
+            return 0
 
 #--------------------------------------------------------------------------
 def insertValueTwitterPosts(tweet_id, user_id, title, retweet_count, favorite_count, created_at, lang, source_url, hashtags):
-  sql = """INSERT INTO twitter_tweets (TID, UID_TWEET, TITLE, RETWEET_COUNT, FAVORITE_COUNT, CREATED_AT, LANGUAGE, SOURCE, HASHTAGS) \
-	VALUES ('%d', '%d', '%s', '%d', '%d', '%s', '%s', '%s', '%s')""" % \
-	(int(tweet_id), int(user_id), title, retweet_count, favorite_count, created_at, lang, source_url, hashtags)
-
-  try:
-    # Execute the SQL command
-    cursor.execute(sql)
-    # Commit your changes in the database
-    db.commit()
-  except:
-    print ("insertValueTwitterPosts:DB connection error!\n")
-    # Rollback in case there is any error
-    db.rollback()
+    if checkEntryExists('twitter_tweets','TID',tweet_id):
+        sql = """UPDATE twitter_tweets SET RETWEET_COUNT = '%d', FAVORITE_COUNT = '%d'
+        WHERE TID = '%d'""" % \
+        (retweet_count, favorite_count,int(tweet_id))
+        try:
+            # Execute the SQL command
+            cursor.execute(sql)
+            # Commit your changes in the database
+            db.commit()
+            return 1
+        except:
+            print ("upadteValueTwitterPosts:DB connection error!\n",sql,"\n")
+            # Rollback in case there is any error
+            db.rollback()
+            return 0
+    else:
+        sql = """INSERT INTO twitter_tweets (TID, UID_TWEET, TITLE, RETWEET_COUNT, FAVORITE_COUNT, CREATED_AT, LANGUAGE, SOURCE, HASHTAGS) \
+        VALUES ('%d', '%d', '%s', '%d', '%d', '%s', '%s', '%s', '%s')""" % \
+        (int(tweet_id), int(user_id), title, retweet_count, favorite_count, created_at, lang, source_url, hashtags)
+        try:
+            # Execute the SQL command
+            cursor.execute(sql)
+            # Commit your changes in the database
+            db.commit()
+            return 1
+        except:
+            print ("insertValueTwitterPosts:DB error!\n",sql,"\n")
+            # Rollback in case there is any error
+            db.rollback()
+            return 0
 
 #--------------------------------------------------------------------------
 
 def RecordTwitterValuesToDB(tweets_list):
     tweet_usr = tweets_list[0].user
-    print("Start insertValueTwitterUsers ",datetime.now().strftime("%H:%M:%S"))
     insertValueTwitterUsers(tweet_usr.id,
-    tweet_usr.name,
+    tweet_usr.name.replace("'", " ").replace('"', ' '),
     tweet_usr.favourites_count,
     tweet_usr.followers_count,
     tweet_usr.friends_count,
     tweet_usr.statuses_count)
-
+    count=0
 
     for i in range (0, len(tweets_list)):
         tweet = tweets_list[i]
@@ -299,16 +416,18 @@ def RecordTwitterValuesToDB(tweets_list):
                 hashtags += tweet.entities['hashtags'][j]['text']
             else:
                 hashtags += tweet.entities['hashtags'][j]['text'] + ', '
-            print("Start insertValueTwitterPosts ",datetime.now().strftime("%H:%M:%S"))
-            insertValueTwitterPosts(tweet.id,
+                
+        count = count + insertValueTwitterPosts(tweet.id,
         tweet.user.id,
-        tweet.full_text,
+        tweet.full_text.replace("'", " ").replace('"', ' '),
         tweet.retweet_count,
         tweet.favorite_count,
         tweet.created_at.strftime("%Y-%m-%d %H:%M:%S"),
         tweet.lang,
         source_url,
         str(hashtags))
+
+    print ('Stored successfully in DB ',count,' tweets')
 
 #--------------------------------------------------------------------------
 
@@ -329,7 +448,7 @@ def twitter_scrapper(twitter_user):
     # fetching the user
     try:
         tweets_list = []
-        for status in tweepy.Cursor(api.user_timeline, id = twitter_user, tweet_mode="extended").items(200):
+        for status in tweepy.Cursor(api.user_timeline, id = twitter_user, tweet_mode="extended").items(600):
             tweets_list.append(status)
     except:
         tweets_list=[]
@@ -360,10 +479,11 @@ except:
     df=[]
 
 # Retrieve tweets
-for i in range (40, len(df)):
+for i in range (38, len(df)):
     twitter_user=df["Twitter"][i]
     print("Start twitter crawl ",datetime.now().strftime("%H:%M:%S"))
     tweets_list = twitter_scrapper(twitter_user)
+    print("Crawled ",len(tweets_list)," tweets")
     
     # Save tweets to DB
 
@@ -373,15 +493,16 @@ for i in range (40, len(df)):
         RecordTwitterValuesToDB(tweets_list)
     except:
         user_id = 0
-        
+
     # Retrieve FB posts
 
     fb_user=df["Facebook"][i]
     print("Start FB crawl ",datetime.now().strftime("%H:%M:%S"))
     post_list = fb_scraper(fb_user)
+    print("Crawled ",len(post_list)," posts")
 
     # Save FB posts to DB
-    
+
     try:
         fb_id=int(post_list[0]['user_id'])
         print("Start Store Posts to DB ",datetime.now().strftime("%H:%M:%S"))
@@ -390,7 +511,7 @@ for i in range (40, len(df)):
         fb_id=0
 
  # Save relation between FB and twitter users
- 
+
     if (user_id & fb_id):
         alias=twitter_user+" - "+fb_user
         print("Start Aliases ",datetime.now().strftime("%H:%M:%S"))
